+++ src/ConfigurationTrait.php
@@ -14,32 +14,65 @@ use ArrayAccess;
 /**
  * ConfigurationTrait which retrieves options from configuration
  *
- * Use this trait if you want to retrieve options from a configuration and optional to perform a mandatory option check
+ * Use this trait if you want to retrieve options from a configuration and optional to perform a mandatory option check.
+ * Default options are merged and overridden of the provided options.
  */
 trait ConfigurationTrait
 {
     /**
-     * @see \Interop\Config\ObtainOptions::options
+     * @see \Interop\Config\RequiresConfig::vendorName
+     */
+    abstract public function vendorName();
+
+    /**
+     * @see \Interop\Config\RequiresConfig::packageName
+     */
+    abstract public function packageName();
+
+    /**
+     * @see \Interop\Config\RequiresConfig::canRetrieveOptions
+     */
+    public function canRetrieveOptions($config)
+    {
+        if (!is_array($config) && !$config instanceof ArrayAccess) {
+            return false;
+        }
+        $id = null;
+
+        if ($this instanceof RequiresContainerId) {
+            $id = $this->containerId();
+        }
+        $vendorName = $this->vendorName();
+        $packageName = $this->packageName();
+
+        if (!isset($config[$vendorName][$packageName])
+            || (null !== $id && !isset($config[$vendorName][$packageName][$id]))
+        ) {
+            return false;
+        }
+        $options = $config[$vendorName][$packageName];
+
+        if (null !== $id) {
+            $options = $options[$id];
+        }
+        return is_array($options) || $options instanceof ArrayAccess;
+    }
+
+    /**
+     * @see \Interop\Config\RequiresConfig::options
      */
     public function options($config)
     {
         if (!is_array($config) && !$config instanceof ArrayAccess) {
             throw new Exception\InvalidArgumentException(
-                sprintf(
-                    '$config parameter provided to "%s" must be an "%s" or "%s"',
-                    __METHOD__,
-                    'array',
-                    'ArrayAccess'
-                )
+                sprintf('Provided parameter $config  must either be of type "array" or implement "\ArrayAccess".')
             );
         }
-
         $id = null;
 
-        if ($this instanceof HasContainerId) {
+        if ($this instanceof RequiresContainerId) {
             $id = $this->containerId();
         }
-
         $vendorName = $this->vendorName();
         $packageName = $this->packageName();
 
@@ -47,23 +80,18 @@ trait ConfigurationTrait
         if (!isset($config[$vendorName][$packageName][$id])) {
             if (!isset($config[$vendorName][$packageName])) {
                 if (empty($config[$vendorName])) {
-                    throw new Exception\RuntimeException(
+                    throw new Exception\OutOfBoundsException(
                         sprintf('No vendor configuration "%s" available', $vendorName)
                     );
                 }
-                throw new Exception\OptionNotFoundException(sprintf(
-                    'No options set in configuration "' . "['%s']['%s']",
-                    $vendorName,
-                    $packageName
-                ));
+                throw new Exception\OptionNotFoundException(
+                    sprintf('No options set for configuration "' . "['%s']['%s']", $vendorName, $packageName)
+                );
             }
             if (null !== $id) {
-                throw new Exception\OptionNotFoundException(sprintf(
-                    'No options set in configuration "' . "['%s']['%s']['%s']",
-                    $vendorName,
-                    $packageName,
-                    $id
-                ));
+                throw new Exception\OptionNotFoundException(
+                    sprintf('No options set for configuration "' . "['%s']['%s']['%s']", $vendorName, $packageName, $id)
+                );
             }
         }
         $options = $config[$vendorName][$packageName];
@@ -71,13 +99,42 @@ trait ConfigurationTrait
         if (null !== $id) {
             $options = $options[$id];
         }
-        // check for mandatory options
-        if ($this instanceof HasMandatoryOptions) {
+        if (!is_array($options) && !$options instanceof ArrayAccess) {
+            throw new Exception\UnexpectedValueException(
+                sprintf(
+                    'Options of configuration ' . "['%s']['%s']%s"
+                    . ' must either be of type "array" or implement "\ArrayAccess".',
+                    $this->vendorName(),
+                    $this->packageName(),
+                    $id ? '["' . $id . '""]' : ''
+                )
+            );
+        }
+        if ($this instanceof RequiresMandatoryOptions) {
             $this->checkMandatoryOptions($this->mandatoryOptions(), $options);
         }
-        // check for default options
-        if ($this instanceof HasDefaultOptions) {
+        if ($this instanceof ProvidesDefaultOptions) {
             $options = array_replace_recursive($this->defaultOptions(), $options);
+        }
+        return $options;
+    }
+
+    /**
+     * Checks if options can be retrieved from config and if not, default options (ProvidesDefaultOptions interface) or
+     * an empty array will be returned.
+     *
+     * @param array|ArrayAccess $config Configuration
+     * @return array|ArrayAccess options Default options or an empty array
+     */
+    public function optionsWithFallback($config)
+    {
+        $options = [];
+
+        if ($this->canRetrieveOptions($config)) {
+            $options = $this->options($config);
+        }
+        if (empty($options) && $this instanceof ProvidesDefaultOptions) {
+            $options = $this->defaultOptions();
         }
         return $options;
     }
@@ -102,7 +159,7 @@ trait ConfigurationTrait
             }
             $id = null;
 
-            if ($this instanceof HasContainerId) {
+            if ($this instanceof RequiresContainerId) {
                 $id = $this->containerId();
             }
 
@@ -111,18 +168,8 @@ trait ConfigurationTrait
                 $mandatoryOption,
                 $this->vendorName(),
                 $this->packageName(),
-                $id ? '[' . $id . ']' : ''
+                $id ? '["' . $id . '""]' : ''
             ));
         }
     }
-
-    /**
-     * @see \Interop\Config\HasConfig::vendorName
-     */
-    abstract public function vendorName();
-
-    /**
-     * @see \Interop\Config\HasConfig::packageName
-     */
-    abstract public function packageName();
 }