+++ test/ConfigurationTraitTest.php
@@ -40,9 +40,59 @@ class ConfigurationTraitTest extends TestCase
     {
         $stub = new ConnectionConfiguration();
 
-        $this->setExpectedException('Interop\Config\Exception\InvalidArgumentException', '$config parameter');
+        $this->setExpectedException('Interop\Config\Exception\InvalidArgumentException', 'Provided parameter');
 
-        return $stub->options('');
+        $stub->options('');
+    }
+
+    /**
+     * Tests canRetrieveOptions()
+     *
+     * @covers \Interop\Config\ConfigurationTrait::canRetrieveOptions
+     */
+    public function testCanRetrieveOptions()
+    {
+        $stub = new ConnectionConfiguration();
+
+        self::assertSame(false, $stub->canRetrieveOptions(''));
+        self::assertSame(false, $stub->canRetrieveOptions(new \stdClass()));
+        self::assertSame(false, $stub->canRetrieveOptions(null));
+
+        self::assertSame(
+            false,
+            $stub->canRetrieveOptions(['doctrine' => ['invalid' => ['default' => ['params' => '']]]])
+        );
+
+        self::assertSame(
+            false,
+            $stub->canRetrieveOptions(['doctrine' => ['connection' => new \stdClass()]])
+        );
+
+        self::assertSame(true, $stub->canRetrieveOptions($this->getTestConfig()));
+    }
+
+    /**
+     * Tests canRetrieveOptions()
+     *
+     * @covers \Interop\Config\ConfigurationTrait::canRetrieveOptions
+     */
+    public function testCanRetrieveOptionsWithContainerId()
+    {
+        $stub = new ConnectionContainerIdConfiguration();
+
+        self::assertSame(false, $stub->canRetrieveOptions(['doctrine' => ['connection' => null]]));
+
+        self::assertSame(
+            false,
+            $stub->canRetrieveOptions(['doctrine' => ['connection' => ['invalid' => ['test' => 1]]]])
+        );
+
+        self::assertSame(
+            false,
+            $stub->canRetrieveOptions(['doctrine' => ['connection' => ['orm_default' => new \stdClass()]]])
+        );
+
+        self::assertSame(true, $stub->canRetrieveOptions($this->getTestConfig()));
     }
 
     /**
@@ -50,13 +100,13 @@ class ConfigurationTraitTest extends TestCase
      *
      * @covers \Interop\Config\ConfigurationTrait::options
      */
-    public function testOptionsThrowsRuntimeExceptionIfNoVendorConfigIsAvailable()
+    public function testOptionsThrowsOutOfBoundsExceptionIfNoVendorConfigIsAvailable()
     {
         $stub = new ConnectionConfiguration();
 
-        $this->setExpectedException('Interop\Config\Exception\RuntimeException', 'No vendor');
+        $this->setExpectedException('Interop\Config\Exception\OutOfBoundsException', 'No vendor');
 
-        return $stub->options(['doctrine' => []]);
+        $stub->options(['doctrine' => []]);
     }
 
     /**
@@ -70,7 +120,7 @@ class ConfigurationTraitTest extends TestCase
 
         $this->setExpectedException('Interop\Config\Exception\OptionNotFoundException', 'No options');
 
-        return $stub->options(['doctrine' => ['connection' => null]]);
+        $stub->options(['doctrine' => ['connection' => null]]);
     }
 
     /**
@@ -78,13 +128,27 @@ class ConfigurationTraitTest extends TestCase
      *
      * @covers \Interop\Config\ConfigurationTrait::options
      */
-    public function testOptionsThrowsOptionNotFoundExceptionIfNoIdOptionIsAvailable()
+    public function testOptionsThrowsOptionNotFoundExceptionIfNoContainerIdOptionIsAvailable()
     {
         $stub = new ConnectionContainerIdConfiguration();
 
         $this->setExpectedException('Interop\Config\Exception\OptionNotFoundException', 'No options');
 
-        return $stub->options(['doctrine' => ['connection' => ['orm_default' => null]]]);
+        $stub->options(['doctrine' => ['connection' => ['orm_default' => null]]]);
+    }
+
+    /**
+     * Tests options() should throw exception if retrieved options not an array or \ArrayAccess
+     *
+     * @covers \Interop\Config\ConfigurationTrait::options
+     */
+    public function testOptionsThrowsUnexpectedValueExceptionIfRetrievedOptionsNotAnArrayOrArrayAccess()
+    {
+        $stub = new ConnectionContainerIdConfiguration();
+
+        $this->setExpectedException('Interop\Config\Exception\UnexpectedValueException', 'Options of configuration');
+
+        $stub->options(['doctrine' => ['connection' => ['orm_default' => new \stdClass()]]]);
     }
 
     /**
@@ -100,8 +164,8 @@ class ConfigurationTraitTest extends TestCase
 
         $options = $stub->options($testConfig);
 
-        $this->assertArrayHasKey('driverClass', $options);
-        $this->assertArrayHasKey('params', $options);
+        self::assertArrayHasKey('driverClass', $options);
+        self::assertArrayHasKey('params', $options);
     }
 
     /**
@@ -117,7 +181,7 @@ class ConfigurationTraitTest extends TestCase
 
         $options = $stub->options($testConfig);
 
-        $this->assertArrayHasKey('orm_default', $options);
+        self::assertArrayHasKey('orm_default', $options);
     }
 
     /**
@@ -136,10 +200,10 @@ class ConfigurationTraitTest extends TestCase
 
         $options = $stub->options($testConfig);
 
-        $this->assertArrayHasKey('params', $options);
-        $this->assertSame($options['params']['host'], $stub->defaultOptions()['params']['host']);
-        $this->assertSame($options['params']['port'], $stub->defaultOptions()['params']['port']);
-        $this->assertSame(
+        self::assertArrayHasKey('params', $options);
+        self::assertSame($options['params']['host'], $stub->defaultOptions()['params']['host']);
+        self::assertSame($options['params']['port'], $stub->defaultOptions()['params']['port']);
+        self::assertSame(
             $options['params']['user'],
             $testConfig['doctrine']['connection']['orm_default']['params']['user']
         );
@@ -151,9 +215,9 @@ class ConfigurationTraitTest extends TestCase
 
         $options = $stub->options($testConfig);
 
-        $this->assertArrayHasKey('params', $options);
-        $this->assertSame($options['params']['host'], $stub->defaultOptions()['params']['host']);
-        $this->assertSame($options['params']['port'], $stub->defaultOptions()['params']['port']);
+        self::assertArrayHasKey('params', $options);
+        self::assertSame($options['params']['host'], $stub->defaultOptions()['params']['host']);
+        self::assertSame($options['params']['port'], $stub->defaultOptions()['params']['port']);
     }
 
     /**
@@ -169,16 +233,16 @@ class ConfigurationTraitTest extends TestCase
 
         $options = $stub->options($testConfig);
 
-        $this->assertArrayHasKey('params', $options);
-        $this->assertSame(
+        self::assertArrayHasKey('params', $options);
+        self::assertSame(
             $options['params']['host'],
             $testConfig['doctrine']['connection']['orm_default']['params']['host']
         );
-        $this->assertSame(
+        self::assertSame(
             $options['params']['port'],
             $testConfig['doctrine']['connection']['orm_default']['params']['port']
         );
-        $this->assertSame(
+        self::assertSame(
             $options['params']['user'],
             $testConfig['doctrine']['connection']['orm_default']['params']['user']
         );
@@ -195,7 +259,7 @@ class ConfigurationTraitTest extends TestCase
         $stub = new ConnectionMandatoryConfiguration();
         $options = $stub->options($this->getTestConfig());
 
-        $this->assertArrayHasKey('orm_default', $options);
+        self::assertArrayHasKey('orm_default', $options);
     }
 
     /**
@@ -209,8 +273,8 @@ class ConfigurationTraitTest extends TestCase
         $stub = new ConnectionMandatoryContainerIdConfiguration();
         $options = $stub->options($this->getTestConfig());
 
-        $this->assertArrayHasKey('driverClass', $options);
-        $this->assertArrayHasKey('params', $options);
+        self::assertArrayHasKey('driverClass', $options);
+        self::assertArrayHasKey('params', $options);
     }
 
     /**
@@ -219,7 +283,7 @@ class ConfigurationTraitTest extends TestCase
      * @covers \Interop\Config\ConfigurationTrait::options
      * @covers \Interop\Config\ConfigurationTrait::checkMandatoryOptions
      */
-    public function testOptionsThrowsRuntimeExceptionIfMandatoryOptionIsMissing()
+    public function testOptionsThrowsMandatoryOptionNotFoundExceptionIfMandatoryOptionIsMissing()
     {
         $stub = new ConnectionMandatoryContainerIdConfiguration();
 
@@ -240,7 +304,7 @@ class ConfigurationTraitTest extends TestCase
      * @covers \Interop\Config\ConfigurationTrait::options
      * @covers \Interop\Config\ConfigurationTrait::checkMandatoryOptions
      */
-    public function testOptionsThrowsRuntimeExceptionIfMandatoryOptionRecursiveIsMissing()
+    public function testOptionsThrowsMandatoryOptionNotFoundExceptionIfMandatoryOptionRecursiveIsMissing()
     {
         $stub = new ConnectionMandatoryRecursiveContainerIdConfiguration();
 
@@ -254,6 +318,40 @@ class ConfigurationTraitTest extends TestCase
         unset($config['doctrine']['connection']['orm_default']['params']['dbname']);
 
         $stub->options($config);
+    }
+
+    /**
+     * Tests options() with recursive mandatory options check
+     *
+     * @covers \Interop\Config\ConfigurationTrait::options
+     * @covers \Interop\Config\ConfigurationTrait::checkMandatoryOptions
+     */
+    public function testOptionsWithRecursiveMandatoryOptionCheck()
+    {
+        $stub = new ConnectionMandatoryRecursiveContainerIdConfiguration();
+
+        $config = $this->getTestConfig();
+
+        self::assertArrayHasKey('params', $stub->options($config));
+    }
+
+    /**
+     * Tests if options() throws a runtime exception if a recursive mandatory option is missing
+     *
+     * @covers \Interop\Config\ConfigurationTrait::optionsWithFallback
+     */
+    public function testoptionsWithFallback()
+    {
+        $stub = new ConnectionDefaultOptionsConfiguration();
+
+        $config = $this->getTestConfig();
+
+        self::assertArrayHasKey('params', $stub->optionsWithFallback([]));
+        self::assertArrayHasKey('params', $stub->optionsWithFallback($config));
+
+        unset($config['doctrine']['connection']['orm_default']['params']);
+
+        self::assertArrayHasKey('params', $stub->optionsWithFallback($config));
     }
 
     /**